rm old MultiProcessManager.py
